# apidoc/swagger.yaml
openapi: 3.0.0
info:
  title: DAZN Movie API
  description: API for managing movies
  version: 1.0.0
paths:
  /movies:
    get:
      summary: Get all movies
      responses:
        "200":
          description: Successful response

    post:
      summary: Add a new movie
      security:
        - BearerAuth: []
      requestBody:
        description: Movie details
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                genre:
                  type: string
                year:
                  type: integer
                streamingLink:
                  type: string
      responses:
        "201":
          description: Movie created successfully
        "401":
          description: Unauthorized

  /movies/{id}:
    put:
      summary: Update a movie
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        description: Updated movie details
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                genre:
                  type: string
                year:
                  type: integer
                streamingLink:
                  type: string
      responses:
        "200":
          description: Movie updated successfully
        "401":
          description: Unauthorized

    delete:
      summary: Delete a movie
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Movie deleted successfully
        "401":
          description: Unauthorized

  /search:
    get:
      summary: Search for movies
      parameters:
        - in: query
          name: q
          schema:
            type: string
      responses:
        "200":
          description: Successful response
  /get-admin-token:
    get:
      summary: Get admin token
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string